config: remove become and clean up takss
Instead of using become, which doesn't make quite sense here, the owner
directive has been used to make it more logical/simple. That was done
for the config, has been expanded to keys, authorized_keys, etc.

Extra paths have been added for the lookup of the ssh keys for the
customer also.